64-bit compability
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1002 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/source/Any.cpp
@@ -496,7 +496,7 @@ int Any::size() const {
     verifyType(ARRAY, TABLE);
     switch (m_type) {
     case TABLE:
-        return m_data->value.t->size();
+        return (int)m_data->value.t->size();
 
     case ARRAY:
         return m_data->value.a->size();
