+++ G3D9/G3D.lib/source/BinaryInput.cpp
@@ -282,7 +282,7 @@ void BinaryInput::loadIntoMemory(int64 startPosition, int64 minLength) {
 #   ifdef G3D_WIN32
         FILE* file = fopen(m_filename.c_str(), "rb");
         debugAssert(file);
-        int ret = fseek(file, (off_t)m_alreadyRead, SEEK_SET);
+        size_t ret = fseek(file, (off_t)m_alreadyRead, SEEK_SET);
         debugAssert(ret == 0);
         size_t toRead = (size_t)G3D::min(m_bufferLength, m_length - m_alreadyRead);
         ret = fread(m_buffer, 1, toRead, file);
